+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 ## Release Notes
 
+### TBD
+
+* Make errors in `SwaggerApi#create` handleable *(Issue #16)*
+
 ### v0.2.0 (2015-08-25)
 
 * Added `Path` object, `SwaggerApi#getPath(reqOrPath)` and `SwaggerApi#getPaths()`